+++ src/components/suggestedUsers/SuggestedUsers.css
@@ -2,7 +2,6 @@
   display: flex;
   flex-direction: column;
   margin-top: 20px;
-  padding: 10px;
 }
 
 .suggestedusers__header {
